+++ test/test_move_first.py
@@ -1,4 +1,3 @@
-from   __future__         import print_function
 import os
 import pytest
 from   in_place           import InPlace
@@ -11,6 +10,7 @@ def test_move_first_nobackup(tmpdir):
     with InPlace(str(p), move_first=True) as fp:
         assert not fp.closed
         for line in fp:
+            assert isinstance(line, str)
             fp.write(line.swapcase())
         assert not fp.closed
     assert fp.closed
@@ -324,7 +324,7 @@ def test_move_first_backup_dirpath(tmpdir):
     not_a_file.mkdir()
     assert pylistdir(not_a_file) == []
     fp = InPlace(str(p), backup=str(not_a_file), move_first=True)
-    fp.write(u'This will be discarded.\n')
+    fp.write('This will be discarded.\n')
     with pytest.raises(EnvironmentError):
         fp.close()
     assert pylistdir(tmpdir) == ['file.txt', 'not-a-file']