+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+v0.2.0 (in development)
+-----------------------
+- Renamed `InPlace` to `InPlaceText` and added a new `InPlace` class for
+  reading & writing `str` objects (whatever those happen to be in the current
+  Python)
+
 v0.1.1 (2017-01-27)
 -------------------
 Rename package & module from "`inplace`" to "`in_place`"  (I could have sworn I